+++ internal/db/database.go
@@ -20,6 +20,7 @@ type database struct {
 	config    config.Config
 	cluster   cluster.Cluster
 	discovery discovery.Client
+	cc        conn.Conn
 }
 
 func (db *database) Discovery() discovery.Client {
@@ -64,6 +65,7 @@ func New(
 	db := &database{
 		config:  c,
 		cluster: cluster.New(ctx, c, pool, c.Balancer()),
+		cc:      pool.Create(endpoint.New(c.Endpoint(), endpoint.WithLocalDC(true))),
 	}
 
 	var cancel context.CancelFunc
@@ -74,17 +76,14 @@ func New(
 	}
 	defer cancel()
 
-	cc := pool.Create(endpoint.New(c.Endpoint(), endpoint.WithLocalDC(true)))
-
 	db.discovery, err = builder.New(
 		ctx,
-		cc,
+		db.cc,
 		db.cluster,
 		db.config.Trace(),
 		opts...,
 	)
 	if err != nil {
-		cc.Release(ctx)
 		return nil, err
 	}