+++ with.go
@@ -10,43 +10,41 @@ import (
 var nextID = uint64(0)
 
 func (c *connection) With(ctx context.Context, opts ...Option) (Connection, error) {
-	if len(opts) == 0 {
-		return c, nil
-	}
+	id := atomic.AddUint64(&nextID, 1)
 
 	opts = append(
 		opts,
 		WithBalancer(
 			c.config.Balancer().Create(),
 		),
-	)
-
-	id := atomic.AddUint64(&nextID, 1)
-
-	opts = append(
-		opts,
 		withOnClose(func(child *connection) {
 			c.childrenMtx.Lock()
 			defer c.childrenMtx.Unlock()
 
 			delete(c.children, id)
 		}),
+		withConnPool(c.pool),
 	)
-	// check if credentials have been overridden
-	tmp := new(connection)
-	for _, o := range opts {
-		_ = o(ctx, tmp)
-	}
-	tmpCfg := config.New(tmp.options...)
-	if tmpCfg.Credentials() == nil {
-		// use previous credentials, so we can share conn pool
-		opts = append(opts, withConnPool(c.pool))
-	}
 
 	child, err := New(
 		ctx,
 		append(
-			c.opts,
+			append(
+				c.opts,
+				With(
+					config.WithSharedPool(), // force set to true prefer shared pool
+					config.WithBalancer(
+						c.config.Balancer().Create(),
+					),
+				),
+				withOnClose(func(child *connection) {
+					c.childrenMtx.Lock()
+					defer c.childrenMtx.Unlock()
+
+					delete(c.children, id)
+				}),
+				withConnPool(c.pool),
+			),
 			opts...,
 		)...,
 	)